% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{calc_do}
\alias{calc_do}
\title{Calculate the mass of dissolved oxgen in epil and hypolimnion(during stratified period)
and the total dissovled oxygen in the lake}
\usage{
calc_do(
  input.values,
  fsed_stratified_epi,
  fsed_stratified_hypo,
  fsed_not_stratified,
  nep_stratified,
  nep_not_stratified,
  min_stratified,
  min_not_stratified,
  wind = NULL,
  khalf = NULL,
  startdate = NULL,
  enddate = NULL
)
}
\arguments{
\item{input.values}{output of the input function}

\item{fsed_stratified_epi}{mg/m3/d}

\item{fsed_stratified_hypo}{mg/m3/d}

\item{fsed_not_stratified}{mg/m3/d}

\item{nep_stratified}{mg/m3/d}

\item{nep_not_stratified}{mg/m3/d}

\item{min_stratified}{mg/m3/d}

\item{min_not_stratified}{mg/m3/d}

\item{wind}{vector of wind speeds?}

\item{khalf}{mg/m3}

\item{startdate}{defaults to NULL}

\item{enddate}{defaults to NULL}
}
\value{
list of datetimes and depths
}
\description{
Calculate the mass of dissolved oxgen in epil and hypolimnion(during stratified period)
and the total dissovled oxygen in the lake
}
