% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{calc_epil_hypo_temp}
\alias{calc_epil_hypo_temp}
\title{Calculate mean epilimnion and hypolimnion surface temperature,using the thermocline depth}
\usage{
calc_epil_hypo_temp(wtemp, td.depth, H)
}
\arguments{
\item{wtemp}{matrix; Water temperatures (rows correspond to time, cols to depth)}

\item{td.depth}{matrix; thermocline depth}

\item{H}{the depth info of the lake}
}
\value{
list of temperatures vector
}
\description{
Calculate mean epilimnion and hypolimnion surface temperature,using the thermocline depth
}
