% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{calc_epil_hypo_vol}
\alias{calc_epil_hypo_vol}
\title{Calculate water epilimnion and hypolimnine volume,using the thermocline depth}
\usage{
calc_epil_hypo_vol(H, A, td.depth, vol_total)
}
\arguments{
\item{H}{vector; the depth info of the lake}

\item{A}{vector; the area info of the lake for each depth}

\item{td.depth}{matrix; thermocline depth}

\item{vol_total}{number;the total volume of the lake}
}
\value{
matrix of epil and hypo volume
}
\description{
Calculate water epilimnion and hypolimnine volume,using the thermocline depth
}
