% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{calc_rmse_epil}
\alias{calc_rmse_epil}
\title{Calculates RMSE for epilimnion}
\usage{
calc_rmse_epil(test_data)
}
\arguments{
\item{test_data}{matrix; Matched observed to simulated data}
}
\value{
double value of RMSE
}
\description{
Calculates RMSE for epilimnion
}
