% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{calc_rmse_hypo}
\alias{calc_rmse_hypo}
\title{Calculates RMSE for hypolimnion}
\usage{
calc_rmse_hypo(test_data)
}
\arguments{
\item{test_data}{matrix; Matched observed to simulated data}
}
\value{
double value of RMSE
}
\description{
Calculates RMSE for hypolimnion
}
