% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{calc_td_depth}
\alias{calc_td_depth}
\title{Calculate thermocline depth}
\usage{
calc_td_depth(wtemp)
}
\arguments{
\item{wtemp}{matrix; Water temperatures (rows correspond to time, cols to depth)}
}
\value{
vector of thermocline depths in m
}
\description{
Calculate planar thermocline depth by checking the highest density gradient over depth.
}
