% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{calc_total_vol}
\alias{calc_total_vol}
\title{Calculate water total volume,using the thermocline depth}
\usage{
calc_total_vol(H, A)
}
\arguments{
\item{wtemp}{matrix; Water temperatures (rows correspond to time, cols to depth)}

\item{td.depth}{matrix; thermocline depth}
}
\value{
list of temperatures
}
\description{
Calculate water total volume,using the thermocline depth
}
