% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{compare_predict_versus_observed}
\alias{compare_predict_versus_observed}
\title{Compares simulated against measured variables}
\usage{
compare_predict_versus_observed(obs, input.values)
}
\arguments{
\item{obs}{matrix; observational data}

\item{input.values}{data frame, simulaedt data}
}
\value{
matched data frame of observed and simulated data
}
\description{
Compares simulated against measured variables
}
