% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{extract_time_space}
\alias{extract_time_space}
\title{Extract time and space information}
\usage{
extract_time_space(wtemp)
}
\arguments{
\item{wtemp}{matrix; Water temperatures (rows correspond to time, cols to depth)}
}
\value{
list of datetimes and depths
}
\description{
Extracts time (from date column) and space (aka depth) information
}
