% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{input}
\alias{input}
\title{Create temperature and volume input values for oxygen model}
\usage{
input(wtemp, H, A)
}
\arguments{
\item{wtemp}{matrix; Water temperatures (rows correspond to time, cols to depth)}

\item{H}{depths}

\item{A}{areas}
}
\value{
vector of thermocline depths in m
}
\description{
Calculate mean epilimnion and hypolimnion surface temperature, as well as volumes.
}
