% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{optim_do}
\alias{optim_do}
\title{Calculates RMSE for total lake}
\usage{
optim_do(
  p,
  input.values,
  fsed_not_stratified = 2e-04,
  nep_not_stratified = 0,
  min_not_stratified = 0,
  wind = NULL,
  proc.obs,
  verbose,
  startdate = NULL,
  enddate = NULL
)
}
\arguments{
\item{p}{not sure what this is}

\item{input.values}{output of the input function}

\item{fsed_not_stratified}{mg/m3/d}

\item{nep_not_stratified}{mg/m3/d}

\item{min_not_stratified}{mg/m3/d}

\item{wind}{vector of wind speeds?}

\item{proc.obs}{dissolved oxygen observations for calibration?}

\item{verbose}{not used}

\item{startdate}{defaults to NULL}

\item{enddate}{defaults to NULL}
}
\value{
double value of RMSE
}
\description{
Calculates RMSE for total lake
}
