% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{valid}
\alias{valid}
\title{check whether the flux is valid and return a valid flux}
\usage{
valid(flux, pool)
}
\arguments{
\item{flux}{to be checked}

\item{pool}{cuurent existing o2 in the lake}
}
\value{
a valid flux
}
\description{
check whether the flux is valid and return a valid flux
}
