% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{weigh_obs}
\alias{weigh_obs}
\title{preprocesses observed data and area-weighs them w/o interpolation}
\usage{
weigh_obs(obs, input.values, H, A)
}
\arguments{
\item{obs}{observed data}

\item{input.values}{input matrix of for instance thermocline depth}

\item{H}{depths}

\item{A}{areas}
}
\value{
matched and weighted-averaged data
}
\description{
preprocesses observed data and area-weighs them w/o interpolation
}
