% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{calc_do}
\alias{calc_do}
\title{Calculate the mass of dissolved oxgen in epil and hypolimnion(during stratified period)
and the total dissovled oxygen in the lake}
\usage{
calc_do(
  input.values,
  fsed_stratified,
  fsed_not_stratified,
  nep_stratified,
  nep_not_stratified,
  min_stratified,
  min_not_stratified,
  wind = NULL
)
}
\arguments{
\item{wtemp}{matrix; Water temperatures (rows correspond to time, cols to depth)}
}
\value{
list of datetimes and depths
}
\description{
Calculate the mass of dissolved oxgen in epil and hypolimnion(during stratified period)
and the total dissovled oxygen in the lake
}
