% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{calc_fit}
\alias{calc_fit}
\title{Calculates RMSE for total lake}
\usage{
calc_fit(input.values, proc.obs)
}
\arguments{
\item{test_data}{matrix; Matched observed to simulated data}
}
\value{
double value of RMSE
}
\description{
Calculates RMSE for total lake
}
