% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{optim_do}
\alias{optim_do}
\title{Calculates RMSE for total lake}
\usage{
optim_do(
  p,
  input.values,
  fsed_not_stratified = 2e-04,
  nep_not_stratified = 0,
  verbose
)
}
\arguments{
\item{test_data}{matrix; Matched observed to simulated data}
}
\value{
double value of RMSE
}
\description{
Calculates RMSE for total lake
}
