% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{preprocess_obs}
\alias{preprocess_obs}
\title{preprocesses observed data and area-weighs them}
\usage{
preprocess_obs(obs, input.values, H, A)
}
\arguments{
\item{obs}{observed data}

\item{H}{depths}

\item{A}{areas}

\item{pool}{input matrix of for instance thermocline depth}
}
\value{
matched and weighted-averaged data
}
\description{
preprocesses observed data and area-weighs them
}
